refactor: rename client file to sanityClient.ts and update code accordingly
