+++ src/pages/Product/Product.tsx
@@ -2,7 +2,7 @@ import React, { ReactNode, useState, useEffect } from 'react';
 import { BsStarFill, BsStarHalf, BsStar } from 'react-icons/bs';
 import NumberFormat from 'react-number-format';
 import { useParams } from 'react-router-dom';
-import sanityClient, { urlFor } from '../../lib/client';
+import sanityClient, { urlFor } from '../../lib/sanityClient';
 import { ProductType } from '../../types';
 // import PRODUCTS from '../../constants/products';
 import Gallery from './Gallery/Gallery';
